+++ integration_test/test/recording_endpoint_test.exs
@@ -86,7 +86,6 @@ defmodule Membrane.RTC.RecordingEndpointTest do
 
     assert File.read!(video_file_path) == File.read!(deserialized_file_path)
 
-    Engine.remove_endpoint(rtc_engine, recording_endpoint_id)
     assert_receive %EndpointRemoved{endpoint_id: ^recording_endpoint_id}
 
     await_report(report_file_path)
@@ -122,7 +121,109 @@ defmodule Membrane.RTC.RecordingEndpointTest do
     assert_receive %TrackAdded{endpoint_id: ^video_file_endpoint_id}, @tracks_added_delay
     assert_receive %TrackAdded{endpoint_id: ^audio_file_endpoint_id}, @tracks_added_delay
     assert_receive %TrackRemoved{endpoint_id: ^video_file_endpoint_id}, @tracks_removed_delay
-    assert_receive %TrackRemoved{endpoint_id: ^audio_file_endpoint_id}
+    assert_receive %TrackRemoved{endpoint_id: ^audio_file_endpoint_id}, @tracks_removed_delay
+
+    assert_receive %EndpointRemoved{endpoint_id: ^recording_endpoint_id}
+
+    await_report(report_file_path)
+    assert %{"tracks" => tracks} = report_file_path |> File.read!() |> Jason.decode!()
+
+    {audio_file, _metadata} =
+      Enum.find(tracks, fn {_filename, track} -> track["type"] == "audio" end)
+
+    {video_file, _metadata} =
+      Enum.find(tracks, fn {_filename, track} -> track["type"] == "video" end)
+
+    audio_deserializer =
+      create_audio_deserializer(%{
+        source: Path.join(output_dir, audio_file),
+        output: deserialized_audio_path,
+        owner: self()
+      })
+
+    video_deserializer =
+      create_video_deserializer(%{
+        source: Path.join(output_dir, video_file),
+        output: deserialized_video_path,
+        owner: self()
+      })
+
+    audio_pipeline = Membrane.Testing.Pipeline.start_link_supervised!(spec: audio_deserializer)
+    video_pipeline = Membrane.Testing.Pipeline.start_link_supervised!(spec: video_deserializer)
+
+    assert_end_of_stream(audio_pipeline, :sink)
+    assert_end_of_stream(video_pipeline, :sink)
+
+    assert deserialized_audio_path |> File.read!() |> byte_size() > 0
+    assert deserialized_video_path |> File.read!() |> byte_size() > 0
+  end
+
+  @tag :tmp_dir
+  test "engine doesn't crash when recording endpoint is removed too early",
+       %{
+         rtc_engine: rtc_engine,
+         tmp_dir: output_dir
+       } do
+    recording_endpoint_id = "recording-endpoint"
+    audio_file_endpoint_id = "audio-file-endpoint"
+    video_file_endpoint_id = "video-file-endpoint"
+
+    audio_file_path = Path.join(@fixtures_dir, "audio.aac")
+    video_file_path = Path.join(@fixtures_dir, "two_seconds.h264")
+
+    recording_endpoint =
+      create_recording_endpoint(rtc_engine, [{Storage.File, %{output_dir: output_dir}}])
+
+    audio_file_endpoint = create_audio_file_endpoint(rtc_engine, audio_file_path)
+    video_file_endpoint = create_video_file_endpoint(rtc_engine, video_file_path)
+
+    :ok = Engine.add_endpoint(rtc_engine, recording_endpoint, id: recording_endpoint_id)
+    :ok = Engine.add_endpoint(rtc_engine, video_file_endpoint, id: video_file_endpoint_id)
+    :ok = Engine.add_endpoint(rtc_engine, audio_file_endpoint, id: audio_file_endpoint_id)
+
+    assert_receive %EndpointAdded{endpoint_id: ^recording_endpoint_id}
+
+    assert_receive %TrackAdded{endpoint_id: ^video_file_endpoint_id}, @tracks_added_delay
+    assert_receive %TrackAdded{endpoint_id: ^audio_file_endpoint_id}, @tracks_added_delay
+
+    Engine.remove_endpoint(rtc_engine, recording_endpoint_id)
+    assert_receive %EndpointRemoved{endpoint_id: ^recording_endpoint_id}
+    refute_receive %EndpointCrashed{endpoint_id: ^recording_endpoint_id}, 1_000
+  end
+
+  @tag :tmp_dir
+  test "creates correct recording, multiple inputs when recording endpoint is removed before all input endpoints finished",
+       %{
+         rtc_engine: rtc_engine,
+         tmp_dir: output_dir
+       } do
+    recording_endpoint_id = "recording-endpoint"
+    audio_file_endpoint_id = "audio-file-endpoint"
+    video_file_endpoint_id = "video-file-endpoint"
+
+    audio_file_path = Path.join(@fixtures_dir, "audio.aac")
+    video_file_path = Path.join(@fixtures_dir, "two_seconds.h264")
+    report_file_path = Path.join(output_dir, @report_filename)
+
+    deserialized_audio_path = Path.join(output_dir, "deserialized_audio.aac")
+    deserialized_video_path = Path.join(output_dir, "deserialized_video.h264")
+
+    recording_endpoint =
+      create_recording_endpoint(rtc_engine, [{Storage.File, %{output_dir: output_dir}}])
+
+    audio_file_endpoint = create_audio_file_endpoint(rtc_engine, audio_file_path)
+    video_file_endpoint = create_video_file_endpoint(rtc_engine, video_file_path)
+
+    :ok = Engine.add_endpoint(rtc_engine, recording_endpoint, id: recording_endpoint_id)
+    :ok = Engine.add_endpoint(rtc_engine, video_file_endpoint, id: video_file_endpoint_id)
+    :ok = Engine.add_endpoint(rtc_engine, audio_file_endpoint, id: audio_file_endpoint_id)
+
+    assert_receive %EndpointAdded{endpoint_id: ^recording_endpoint_id}
+
+    assert_receive %TrackAdded{endpoint_id: ^video_file_endpoint_id}, @tracks_added_delay
+    assert_receive %TrackAdded{endpoint_id: ^audio_file_endpoint_id}, @tracks_added_delay
+    assert_receive %TrackRemoved{endpoint_id: ^video_file_endpoint_id}, @tracks_removed_delay
+    refute_received %TrackRemoved{endpoint_id: ^audio_file_endpoint_id}
 
     Engine.remove_endpoint(rtc_engine, recording_endpoint_id)
     assert_receive %EndpointRemoved{endpoint_id: ^recording_endpoint_id}
@@ -203,7 +304,6 @@ defmodule Membrane.RTC.RecordingEndpointTest do
     assert_receive %TrackRemoved{endpoint_id: ^video_file_endpoint_id2}, @tracks_removed_delay
     assert_receive %TrackRemoved{endpoint_id: ^audio_file_endpoint_id2}, @tracks_removed_delay
 
-    Engine.remove_endpoint(rtc_engine, recording_endpoint_id)
     assert_receive %EndpointRemoved{endpoint_id: ^recording_endpoint_id}
 
     await_report(report_file_path)
@@ -274,7 +374,6 @@ defmodule Membrane.RTC.RecordingEndpointTest do
     assert_receive %TrackAdded{endpoint_id: ^video_file_endpoint_id}, @tracks_added_delay
     assert_receive %TrackRemoved{endpoint_id: ^video_file_endpoint_id}, @tracks_removed_delay
 
-    Engine.remove_endpoint(rtc_engine, recording_endpoint_id)
     assert_receive %EndpointRemoved{endpoint_id: ^recording_endpoint_id}
 
     assert_receive :upload_initialized
@@ -401,7 +500,6 @@ defmodule Membrane.RTC.RecordingEndpointTest do
 
       refute_received %EndpointCrashed{endpoint_id: ^recording_endpoint_id}
 
-      Engine.remove_endpoint(rtc_engine, recording_endpoint_id)
       assert_receive %EndpointRemoved{endpoint_id: ^recording_endpoint_id}
     end